+++ lib/src/basic/types/item.dart
@@ -325,7 +325,6 @@ Map<String, dynamic> _copyDeepMap(Map<String, dynamic>? map) {
   final newMap = <String, dynamic>{};
 
   map?.forEach((key, value) {
-    print(value);
     newMap[key] =
         (value is Map) ? _copyDeepMap(value.cast<String, dynamic>()) : value;
   });